fix: get rid of render module and update imports
